Better implementation of reset app

+++ framework/storage/src/main/java/nl/rijksoverheid/mgo/framework/storage/file/EncryptedFileStore.kt
@@ -77,6 +77,8 @@ internal class EncryptedFileStore(
 
     override suspend fun deleteFile(name: String) {
         val file = File(dir, name)
-        check(file.delete()) { "Could not delete file" }
+        if (file.exists()) {
+            check(file.delete()) { "Could not delete file" }
+        }
     }
 }
